no comment

+++ tasks/omp/Sadikov_I_SparseMatrixMultiplication_OMP/perf_tests/main.cpp
@@ -20,7 +20,7 @@ std::vector<double> GetRandomMatrix(int size) {
   int high = 500;
   int low = 0;
   std::uniform_int_distribution<> number(low, high);
-  for (int i = 0; i < (static_cast<int>(size / 5)); ++i) {
+  for (int i = 0; i < size / 5; ++i) {
     data[i] = static_cast<double>(number(gen));
   }
   std::ranges::shuffle(data, gen);
